Tighten bounds clarity in optimized PDB segment path
Agent-Logs-Url: https://github.com/khb7840/foldcomp/sessions/ab19eb9a-a1a8-4a67-817d-382fbf18fc7b

+++ src/atom_coordinate.cpp
@@ -448,8 +448,8 @@ void writeAtomCoordinatesToPDB(
         appendTitleLines(output, title);
     }
 
-    size_t total = atoms.size();
-    for (size_t i = 0; i < total; i++) {
+    int total = static_cast<int>(atoms.size());
+    for (int i = 0; i < total; i++) {
         const AtomCoordinate& atom = atoms[i];
         output.append("ATOM  ", 6);
         appendRightAlignedInt(output, atom.atom_index, 5);
